+++ fl_image_picker/lib/src/asset_builder.dart
@@ -1,4 +1,4 @@
-import 'package:fl_assets_picker/fl_assets_picker.dart';
+import 'package:fl_image_picker/fl_image_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';